Make URL from SOM bugs more useful
As in the bug, we are not showing enough information when filing a bug, so the users do not know which build has been failing.

I added the first failure URL, so they will know what has failed. Also change the text a bit to make it not so clustered.

Here is what it looks like https://screenshot.googleplex.com/qH1Qi39J92q

R=jsca@google.com

Bug: 981013
Change-Id: Ia6c339fb7d04c731fee933ed278d6a71a51354d8
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1835313
Commit-Queue: Quang Minh Tuan Nguyen <nqmtuan@google.com>
Reviewed-by: Joey Scarr <jsca@chromium.org>
Cr-Commit-Position: refs/heads/master@{#26243}