[Findit V2] Fix bad field mask when retrieving rerun dimensions.
'infra.swarming.taskDimensions' does not work as a FieldMask by the
buildbucket v2 client in Findit (prpc error 3),
although it does work when creating a request on rpcexplorer directly.

To make this work I have fixed the casing of the last element in the
path to 'task_dimensions'.

R=chanli,nodir

Bug: 1012855
Change-Id: I3ea4d6f7e3d9810f7fad728ff95b9a89084c4903
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1850249
Reviewed-by: Chan Li <chanli@chromium.org>
Reviewed-by: Nodir Turakulov <nodir@chromium.org>
Commit-Queue: Roberto Carrillo <robertocn@chromium.org>
Cr-Commit-Position: refs/heads/master@{#26384}