[Findit V2] Whitelist chromium CI builders for compile analysis.
This list should not be maintained as a fixed configuration in source
code, instead we should derive it from buildbucket or swarmbucket. But
for the time being, in order to validate the compile failure analysis
implementation in Findit V2 (and later auto-actions) we need to provide
reasonable coverage of the CI builders.

R=chanli

Bug: 1012485
Change-Id: Id4d878d33de7815d333869d289ad6ed360975533
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1848396
Reviewed-by: Chan Li <chanli@chromium.org>
Commit-Queue: Roberto Carrillo <robertocn@chromium.org>
Cr-Commit-Position: refs/heads/master@{#26388}