[code coverage] Make allowed gitiles host configurable
Main motivation is to unhard chromium/chrome bits and onboard other
projects with as little efforts as possible.

The allowed gitiles host is used as an example, and once everyone
agrees on the direction to move forward, and make other hard-codings
configurable the same way in next CLs.

Bug: 977639
Change-Id: I314cd5bf334d9b27184836dac9fd02d4fb38e960
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1862668
Reviewed-by: Nodir Turakulov <nodir@chromium.org>
Reviewed-by: Sajjad Mirza <sajjadm@chromium.org>
Commit-Queue: Yuke Liao <liaoyuke@chromium.org>
Cr-Commit-Position: refs/heads/master@{#26494}