[som] Use the 'chrome' project when querying for chromium and android alerts.
The 'chromium' buildbucket tables only contain entries for builds in the
'chromium' project. The 'chrome' tables contain entries for builds in both
the 'chromium' and 'chrome' projects. (See crbug.com/1015209.)

Since the chromium and android trees in SoM are configured to alert on
builds in both of those projects, this switches their analyzers to look
at the 'chrome' project views.

Bug: 938113
Change-Id: I2b02854aaaf9c7d02def28a5d7e4c178f808b407
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1866071
Reviewed-by: Quang Minh Tuan Nguyen <nqmtuan@google.com>
Reviewed-by: Joey Scarr <jsca@chromium.org>
Commit-Queue: Ben Pastene <bpastene@chromium.org>
Cr-Commit-Position: refs/heads/master@{#26542}