[Monorail] Fix merging of issue starrers.
The refactoring of logic from issuedetailezt.py to work_env.py missed
the needed call to merge starrers when a user merges issues.  So, this
CL adds that logic back.

It may have been broken for longer than that because the updated
star count was not on on the copy of the issue that was written
on the next step in the logic (merging CCs).  So, the issue count
would have been updated correctly in the DB and then immediately
set back to an incorrect value.  There were unit tests that covered
that, but fake.py is not realistic enough to catch the stale data
problem.

R=jojwang@chromium.org

Bug: monorail:6465
Change-Id: I3e03e96aa570b825e4d34b5c714f48202b8ef9b5
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1863995
Reviewed-by: Joanna Wang <jojwang@chromium.org>
Commit-Queue: Jason Robbins <jrobbins@chromium.org>
Cr-Commit-Position: refs/heads/master@{#26622}