[Findit V2] Implement auto-actions for compile culprits.
This CL implements the main flow of the actions to be taken when a
culprit is found. It adds the OnCulpritFound API to analysis_api and
implements it for CompileAnalysisAPI subclass.

It also adds the necessary configurations, and misc changes.

See go/finditv2-compile-actions for details

R=chanli,nodir,orodley

Bug: 1008995
Change-Id: I56866671ebdf43e47e8f0d58d0eb7b1d5c819f64
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1873011
Commit-Queue: Roberto Carrillo <robertocn@chromium.org>
Reviewed-by: Chan Li <chanli@chromium.org>
Cr-Commit-Position: refs/heads/master@{#26668}