[Monorail] Add pubsub task queue for publishing issue updates
[Important] For deploying this CL - queue.yaml needs to be
updated manually.

This CL adds a new task to push a message onto a pub/sub queue
every time an issue is updated. The only information in the
message is the issue ID. Any further additions to data in the
message should be security reviewed.

Based on parts of https://crrev/c/1636689 and its ancestors.

Bug: monorail:5440
Change-Id: Ieef73a4bcdf82cd8703b5a7f90ed7ddafe281901
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1688276
Reviewed-by: Joanna Wang <jojwang@chromium.org>
Commit-Queue: Jeff Carpenter <jeffcarp@chromium.org>
Cr-Commit-Position: refs/heads/master@{#26675}