[Monorail] Delete obsolete benchmark directory.
This was used for load testing the EZT issue list page back in 2015,
and at one point it had some code for a faster way to parse responses
from an early version of Cloud SQL.  Neither of those has been needed
in a long time.  We have lots of real data under real daily load, so
I don't expect this code to be needed again, and if it was, it would
probably be done with a different tool.

R=jojwang@chromium.org

Change-Id: Idf62126fa95481f66b09d99b67c9c7d2495766b2
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1875477
Reviewed-by: Joanna Wang <jojwang@chromium.org>
Commit-Queue: Jason Robbins <jrobbins@chromium.org>
Cr-Commit-Position: refs/heads/master@{#26695}