Use bug_list instead of bugs when logging event
The table sheriff-o-matic.events.annotations is broken. The bugs column should be a struct instead of string. It made all logging related to bugs result in a failure

As changing the data type from string -> struct is quite troublesome and we may end up corrupt the data, let's create a new column for it, name "bug_list"

Later when we think it is safe or we don't need the old data, we can just drop the "bugs" column altogether.

Tested with staging https://screenshot.googleplex.com/rJygyaWrD4o



Bug: 1018635
Change-Id: I8fe484b08399ba1593c6989c1b12c40c22778423
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1885354
Commit-Queue: Quang Minh Tuan Nguyen <nqmtuan@google.com>
Reviewed-by: Joey Scarr <jsca@chromium.org>
Cr-Commit-Position: refs/heads/master@{#26818}