[Monorail] Include attachment_count in ListIssues when database value is positive.
Tested with test.py and verified with a manually corrupted database to ensure the results returned to the frontend were sane.

The attachment_count field is left unset rather than setting to 0 so that users of the API can tell when their query has returned an accurate attachment count.

R=jessan@google.com, jojwang@chromium.org

Bug: monorail:5857
Change-Id: Iad2c1563a0aeef9a49cff06dcf3e9e887c452afb
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1885604
Commit-Queue: Philip Walls <pawalls@google.com>
Reviewed-by: Joanna Wang <jojwang@chromium.org>
Cr-Commit-Position: refs/heads/master@{#26845}