[som][fuchsia] Support custom filter for bigquery_analyzer.
This cl adds a new setting to the admin page that allow users to set the build bucket project for a specific tree without having to alter the SQL template in bigquery_analyzer.go.

For example, if fuchsia tree wants to only look for a specific build bucket project for build copping, we can just update the admin page instead of checking a new cl.

For more advanced query changes, a cl is still required.


Bug: fxb/37615
Change-Id: I19d405a210e1bddfe83d8440525cadce26e0ce6d
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1881623
Commit-Queue: Yuan Zhi <yuanzhi@google.com>
Reviewed-by: Joey Scarr <jsca@chromium.org>
Reviewed-by: Quang Minh Tuan Nguyen <nqmtuan@google.com>
Cr-Commit-Position: refs/heads/master@{#26860}