Fix empty string project ID when filing bugs
We got quite a few bugs with ProjectID = "" in the DB, it leads to the failures in our refresh annotation cron job.

It looks like this happens when we link a bug to an alert with the form "crbug.com/123456". The project ID inserted is "", instead of "chromium".

We actually have tests for this, but just that we got the wrong assert. So I fixed all the tests in this CL as well.

Bug: 1018635

Change-Id: Ibbacfd575367264e45809e3954ba01862dd5d868
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1894771
Reviewed-by: Owen Rodley <orodley@chromium.org>
Commit-Queue: Quang Minh Tuan Nguyen <nqmtuan@google.com>
Cr-Commit-Position: refs/heads/master@{#26876}