[3pp/cpython] Prevent picking up broken vpython when running python_mod_gen.
This is fixed already in cpython3, but didn't get backported to cpython2.

When we're not cross-compiling, it's important to use the interpreter we
built because the '-S' flag goofs up vpython-native... vpython used to drop
the '-S' flag, but a well-meaning contributor removed this behavior.

R=tandrii@chromium.org, vadimsh@chromium.org

Bug: 1013097
Change-Id: I2b28ac8a6abb5586469bd18030778f413766f1d6
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1906535
Auto-Submit: Robbie Iannucci <iannucci@chromium.org>
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Reviewed-by: Marc-Antoine Ruel <maruel@chromium.org>
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>
Cr-Commit-Position: refs/heads/master@{#26972}