qscheduler: add a gauge metric to track utilization rate by label
In order to determine hotspot labels, it would be helpful for
quotascheduler to emit a per-label metric indicating the number of
running and idle bots and tasks that possess that label.

This is a little tricky to do properly with gauge metrics, hence the
creation of a new labelUtilizationCounter utility class that keeps an
in-memory cache of all previously encountered labels.

BUG=1025935
TEST=None

Change-Id: I1885a8aeae0b9e9597a8d7b9268c1d88ec9c8b23
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1922549
Reviewed-by: Prathmesh Prabhu <pprabhu@chromium.org>
Commit-Queue: Aviv Keshet <akeshet@chromium.org>
Cr-Commit-Position: refs/heads/master@{#27207}