Remove gae_ts_mon from chromium_rietveld
The DjangoMiddleware version of gae_ts_mon that Rietveld
uses has been broken for quite some time (no metrics
reported in the last year). This application is sufficiently
non-critical that it is more appropriate to remove this
defunct monitoring than it is to try to fix it.

Bug: 1003185
Change-Id: I6740196e507946a8ef1766f978abd2eb0294548b
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1928645
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Commit-Queue: Aaron Gable <agable@chromium.org>
Cr-Commit-Position: refs/heads/master@{#27275}