[3pp/cpython] Get cpython building correctly on all platforms.
I hope :).

This works for native compiles (so, linux-amd64 and mac-amd64) by first
building a 'bootstrap' python interpreter and installing it into the
checkout. This gives us an interpreter which is good enough to run
setup.py later. Then we reconfigure and remake python, including the
full static treatment for all known modules.

We could formalize this for cross-compiles by requiring cpython as a
tool for cross compilation targets... but this is very likely the LAST
build of python2 we'll need (again, I hope), so I'm inclined to just
leave it be.

TBR=tandrii@chromium.org

Bug: 1026340
Change-Id: Ib239400821106d7793a57bd9ef25a63e0bf09547
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1932217
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>
Cr-Commit-Position: refs/heads/master@{#27301}