Allow deleting comments
Sheriffs are getting a lot of noises because of old unrelated comments. Worst, they don't have a way to delete the comments, and they will follow the alert forever.

At least we should provide a way for the to delete comments that are no longer relevant.

Actually we are supporting this features, just that we only allow the owner of a comment to remove it.

While this model makes senses in some cases, in this case, it does not, because of the rotational nature of sheriffing.

Instead, I think we should just confirm with the users if they want to delete a comment or not, and trust them to do the right job.

In the worst case, we have the eventlog table in our DB, so we will always know who did what.

Bug: 894980
Change-Id: I329a83cee5898fbc9cc66f8f844e3a7c4972ae47
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1935687
Commit-Queue: Quang Minh Tuan Nguyen <nqmtuan@google.com>
Reviewed-by: Owen Rodley <orodley@chromium.org>
Cr-Commit-Position: refs/heads/master@{#27341}