Check in Dockerfile packages for linux-arm64 swarming bots.
The build.sh file checks its host operating system. If it is an arm64 system, arm64-specific packages are added to the docker image.  The arm64 packages are required to run Cr-Fuchsia qemu-kvm tests in arm64 swarming bots.  If if the host is not an arm64 system, no packages are added to the docker image.

Bug: 1028320
Change-Id: Ibef6253741a2bf02d6c32e12a6fcb4c20563127a
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1933034
Commit-Queue: Stephen Roe <steveroe@google.com>
Reviewed-by: Elliott Friedman <friedman@chromium.org>
Cr-Commit-Position: refs/heads/master@{#27354}