[recipes] Replace cipd.platform_suffix method usage with package variables
This is part of the effort to remove the depot_tools' cipd recipe module and
substitute it with the one in recipe engine. The cipd module in recipe engine
no longer has the platform_suffix method and using package variables is the
recommended replacement.

R=iannucci

Bug: 875523
Change-Id: Ie6387b5a32ae9229bc430ae488405b3d3a633afa
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1949690
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Commit-Queue: Yiwei Zhang <yiwzhang@google.com>
Cr-Commit-Position: refs/heads/master@{#27610}