Have recipe_roll_tryjob rebase CL
This avoids false positives in the case where the parent of the CL being
tested is stale, and there have been nontrivial commits to master since
the CL was last rebased. Previously, such a CL would be marked as
nontrivial because it would appear to revert the nontrivial master
commits, and a revert of a nontrivial commit is generally nontrivial
itself. We can avoid this by rebasing the CL onto the latest master
commit before training the downstream repo against the CL.

Bug: 1023819
Bug: fuchsia:37585
Change-Id: Ide187884004f164c4d44251cad1b613f43f26e33
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1955900
Commit-Queue: Oliver Newman <olivernewman@google.com>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Cr-Commit-Position: refs/heads/master@{#27706}