[arquebus] perform duplicate checks for Assignee and CcRefs with UserRef.DisplayName
Arquebus creates UserRef{} the email address from UserSource_Email or the one
retrieved from RotaNG. For both types of UserSource{}s, Arquebus creates
UserRef{} with the email addresses in DisplayName.

However, Arquebus doesn't invoke monorail.GetUser() to validate the email
addresses, and, as a result, the UserID of the UserRef{}s created from Assigner
configs are always 0.

createDelta() invokes proto.Equal() to find if a given Issue already has
the intended owner assigned, and builds map[UserRef.UserId] to create a list
of UserRefs[] to be newly added into the CC list. However, both checks produce
false results, because monorail.GetIssue() returns an Issue{} object with
the UserID in Issue.OwnerRef{} and Issue.CcRefs{}.

This CL changes the checks such that createDelta() performs checks based on
UserRef.DisplayName, which is the email address.

I manually reverted the patch on issue.go, and ran the new tests to verify that
the new tests fail without the patch.


R=tandrii

Bug: 1032306
Change-Id: I775429c71014b08bbaf2c04d8c89054113d49aa8
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1958549
Commit-Queue: Scott Lee <ddoman@chromium.org>
Reviewed-by: Andrii Shyshkalov <tandrii@google.com>
Cr-Commit-Position: refs/heads/master@{#27739}