[recipes/infra_system] Add option to disable crazy infra env.
This doesn't exist on the arm64 bots (which are presently docker
containers). That said, we should NEVER need to deploy infra_python on
linux-arm64. Or anything new...

I also fixed the trybot version, as it was already wrong
compared to the ci bots; On CI the infra system env wraps
gclient_runhooks, causing "ENV" to generate with the global python
installation. In the trybots it's generating ENV with the per-task
python, THEN changing the python and doing the cipd pkg/test cycle.
The global python has NO effect on the pkg/test steps.

R=tandrii@chromium.org, vadimsh@chromium.org

Bug: 1026340
Change-Id: I4855e06d19926718dc3ae7d1ecb7720d03e07e30
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1959746
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Auto-Submit: Robbie Iannucci <iannucci@chromium.org>
Cr-Commit-Position: refs/heads/master@{#27760}