[gae_ts_mon] reland "add instrument_flask" with additional import patches
https://crrev.com/c/1958284 was reverted at https://crrev.com/c/1966311 because
it requires flask listed in the app.yaml as a dependency, and raises
an exception with an import error, otherwise.

This CL contains the same changes of https://crrev.com/c/1958284, but additional
changes, at config.py, to handle the import error.

-----------------------------------------------------------------------
Note that unit tests are not able to catch this error, because unit tests run
in a virtual environment,
 - that is built by https://cs.chromium.org/chromium/infra/bootstrap/, and
 - where flask is already available.

Therefore, I launched a devserver instance, with "make serve", to see if the
same import error occurs or not.
: https://screenshot.googleplex.com/gcozcfgKH7G
: https://screenshot.googleplex.com/ywWQ0GbGdCm

I was only able to visit the main page, and am not sure if I have to test more,
such creating a project, registering a new user, filing a new issue, etc.
Please let me know if so.

Also, during my testing, I experienced an import failure for utils.py, even with
the patches included in this CL. I didn't finish all the steps, listed in
README.md, for setting a local devserver instance, because the import failure
already occurs in devserver when I run "make serve".

I played with vpython and sys.path to find the cause of the import failure, and
checked out an old revision to find since what revision the import failure
started occurring. However, none of the investigations gave me a clear idea
about what the cause is.

After that, I followed all the reamining steps to setup Monorail dev, and,
surprisingly, the import failures for utils.py stopped occurring, I still have
no idea what the cause of the import failure for utils.py, and will happen again
or not.

Unfortunately, the existing unit test infrasturure doesn't help much, as the
tests run in a different environment. Please let me know if there is a better
way to check if the same issue still exists in Monorail, before this CL
submitted.

R=iannucci@chromium.org,kweng@google.com,dtu@chromium.org

Bug: 1003185
Change-Id: Ia5b10b0540a375ece6a858420e93f0cdab5e9348
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1966190
Reviewed-by: Kevin Weng <kweng@google.com>
Reviewed-by: Dave Tu <dtu@chromium.org>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Commit-Queue: Scott Lee <ddoman@chromium.org>
Cr-Commit-Position: refs/heads/master@{#27900}