[monorail] Move projectName from an <mr-app> property to the Redux store.
We need the projectName from the URL in the Redux store so
that the selectors know which project's data to look up.

There's already a projectName selector that gets it from the project
config, we'll probably want to get rid of it in a follow-up.

Bug: chromium:5953
Change-Id: I2f4f0c34ea740d949c8ec0b273134769a2cab8c8
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1968265
Reviewed-by: Tiffany Zhang <zhangtiff@chromium.org>
Commit-Queue: Dave Tu <dtu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#27921}