BigQuery helper bqh.send_rows should not mutate input rows.
This leads to very surprising results if the call is retried
with seemingly same input. This is only a problem if protobuf messages
are sent as rows, since these are modified to dicts, which are then
rejected during retries.

R=qyearsley

Bug: 804289
Change-Id: If135810b65f3ba5ae375405fa95a28c916550bf6
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1979465
Commit-Queue: Andrii Shyshkalov <tandrii@google.com>
Auto-Submit: Andrii Shyshkalov <tandrii@google.com>
Reviewed-by: Quinten Yearsley <qyearsley@chromium.org>
Cr-Commit-Position: refs/heads/master@{#27994}