Monorail: Fix bug where issue loading snackbar would get double dispatched when navigating from an issue detail page back to the issue list page.
To reproduce this regression before this change:

- Go to the new list view.
- Click on any issue.
- Click back.
- Two snackbars saying "Loading issues..." will show up. Only one is expected.

The cause of this bug:
(This was a difficult bug to track down. I asked about the cause of this bug in the Web Components Users chat, and I may want to rethink our best practices around how to use update() vs. updated() based on this bug.)

- this.fetchingIssueList starts off as undefined because it's an unset.
- this.fetchingIssueList is initialized to false in the constructor and stateChanged() also sets it to false. The initial update() for the element is triggered.
- update() shows a change from undefined to false. This is correct.
- stateChanged() is called, which sets boolean to true. BUT this stateChanged() call happens between update() and the fetchingIssueList check in updated().
- updated() shows old value of boolean as undefined and new value as true (not false).
- update() is called again showing old value as false and new value as true because of the last stateChanged().
- updated() called again with false -> true as well.

Basically, there's a chain where the value goes from undefined -> false -> true, which is two updates, but updated() catches two changes from undefined -> true and false -> true because the second change happens in between when update() and updated() are called, causing the final value in the first updated() call to be incorrect.

How this change fixes the bug:

- By moving this logic into the beginning of updated(), we run the check before any other stateChanged() callbacks can change the value of fetchingIssueList.

Testing:

- I tried, but I wasn't able to find a way to reproduce this issue with tests, unfortunately.

Bug: monorail:5915
Change-Id: If261823c29f84ac155e16478c5ce7bee22d9978e
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1982694
Commit-Queue: Tiffany Zhang <zhangtiff@chromium.org>
Reviewed-by: Joanna Wang <jojwang@chromium.org>
Cr-Commit-Position: refs/heads/master@{#28021}