[tricium] Validate docstring params
Enables the docparams pylint plugin. If a docstring contains an "Args:"
section, docparams validates the section contains every argument, and
that each argument has a type in parentheses. This will help avoid
docstrings getting out-of-date with the corresponding function
signatures.

If docstring warnings become too annoying, we can enable this just for
Fuchsia in the future.

Bug: fuchsia:43524
Change-Id: I6ca98d9a863c6ff45289ec5600a9931e84ba95e0
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1987492
Reviewed-by: Marc-Antoine Ruel <maruel@chromium.org>
Reviewed-by: Quinten Yearsley <qyearsley@chromium.org>
Commit-Queue: Oliver Newman <olivernewman@google.com>
Cr-Commit-Position: refs/heads/master@{#28143}