Monorail: Create initial design for project list page using card layout with mock data.
Demo here: https://28140-e966bbb-tainted-zhangtiff-projects-dot-monorail-dev.appspot.com/projects

This change is essentially an HTML + CSS mock of a proposed design for a new project list page on Monorail. The mock uses fake data but could easily be changed to use real data.

This change also lays some basic scaffolding for a new projects page, but the actual design of the projects page could be changed easily if we don't like it.

For feedback on this change I'd like to focus on receiving:

* Visual design feedback. Feel free to nit the white space, typography, or whatever. In particular, if we like the general styling on the page, I'd like to look into changing more of the typography on Monorail to look like this.
* A general thumbs up/down on the idea of replacing the home page of Monorail with cards rather than a table.


Bug: monorail:6966
Change-Id: I1e5c004d0c43aa68741986c58b287aff9632ffc6
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1986171
Commit-Queue: Tiffany Zhang <zhangtiff@chromium.org>
Reviewed-by: Dave Tu <dtu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#28165}