Get build criticality from latest build instead of step_status_transition
There are a few cases that we show non-critical chromeOS builders (we should not show).

It looks like the problem is that we are getting the criticality information from step_status_transitions views, which will look at the past 7 days. If in the past 7 days, all builds for a builder failed at the same step, the builder will not be included in the view, so no criticality information can be retrieved.

Instead, we should get the criticality information from the top failure instead.

Tested in staging.

Bug: 1002965
Change-Id: I26e594575873bbe14c907b822abe34c2840e8637
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1994752
Reviewed-by: Owen Rodley <orodley@chromium.org>
Commit-Queue: Owen Rodley <orodley@chromium.org>
Cr-Commit-Position: refs/heads/master@{#28277}