cros lab inventory: don't return error when requests are fulfilled
Though the operation on some devices may fail, the request itself is
good, so it make sense to set err as nil.
Additionally, the request is for a batch of devices, there might be
various category of errors, so it's hard to say what's the error in
general.

BUG=chromium:962689
TEST=make test

Change-Id: I9244ccac0a4a98da8097b07d52f729ed032f38e8
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2001191
Commit-Queue: Congbin Guo <guocb@chromium.org>
Auto-Submit: Congbin Guo <guocb@chromium.org>
Reviewed-by: Xixuan Wu <xixuan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#28342}