[Flake] Skip processing non-hidden flake data if there are too many
When a CL has a bug, it sometimes causes thousands of tests to randomly
fail, which results in hundreds of tests to be considered as flaky,
and it introduces noises to Flake Portal because these failures are not
really due to flaky tests.

This CL reduces the noise by skip processing flake data if there are
too many failures.

Bug: 1042266
Change-Id: Id4d3eac7bbbad110ef8d4789bec59559aeb3c22a
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2003768
Commit-Queue: Yuke Liao <liaoyuke@chromium.org>
Reviewed-by: Chan Li <chanli@chromium.org>
Cr-Commit-Position: refs/heads/master@{#28365}