[monorail] Make mr-show-columns-dropdown only include columns in the shown issues.
This is a simplification of the issue list column selector logic
so that it's populated from the shown issues rather than from the
selected project config.  This is easier for hotlists, which have
no selected projects, and allows us to skip fetching the project
configs for all the issues.

This changes the items shown in the dropdown for both the main issue
list and hotlist issue list.

Pros:
* Code is a little bit simpler.
* More performant for hotlists, since we skip fetching project configs.
* Items irrelevant to the current view are excluded,
  which makes relevant items easier to find.
* Ad-hoc labels are now included in the list.

* Users looking for an unset field won't be able to find it.
* If a field isn't set in the first page of issues, it won't be in the list.

Cons: 
Change-Id: I1d2e30e429e93a630198acc97e712a220f75cf18
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2032030
Reviewed-by: Kevin Weng <kweng@google.com>
Reviewed-by: Tiffany Zhang <zhangtiff@chromium.org>
Commit-Queue: Dave Tu <dtu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#28728}