[monorail] Remove karma-spec-reporter.
Monorail has 4 test reporters defined:
  mocha, spec, chopsui-json, and coverage.

chopsui-json is used on the bots to output JSON
in tests-results.appspot.com-compatible format.

coverage is used to output coverage information.

mocha and spec are both used to output the list of tests and test
failure details to stdout. So they are redundant with each other,
causing the test results to be outputted twice (out of sync with
each other).

I've chosen mocha because it has a slightly better output format for
test failures, but it does have a weird issue: it outputs the test
list when run on a single shard, but doesn't output the test list
when run in parallel. Maybe this is the behavior we want anyway:
in debug mode you are single-sharded and might want to run a subset of
tests, so it's useful to see the test list. But in prod mode you want
to run all tests and just the summary with the total number of tests
and the failures is more useful.

The mocha, spec, and chopsui-json reporters also all output
console.log() messages to the terminal, so this reduces
the number of redundant messages from 3 to 2. We probably
want to also update chopsui-json to stop doing that.

Change-Id: I52db424244f22da33954883a8528d8a83c1fdf63
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2033543
Commit-Queue: Dave Tu <dtu@chromium.org>
Auto-Submit: Dave Tu <dtu@chromium.org>
Reviewed-by: Tiffany Zhang <zhangtiff@chromium.org>
Cr-Commit-Position: refs/heads/master@{#28756}