Stop munging shift dates in the rota-ng client.
Before this change, rendering the shift list would modify the dates,
converting them from a ISO 8601 string to a Luxon DateTime object. These
dates would then get converted back to strings and sent to the server in
the update request.

After this change, shift dates are left in ISO 8601 form in the data
model and converted only for rendering (not in-place like before).

Bug: 1045022
Change-Id: Ib4f190651cb5065be314384d95d47d0494c2b0e8
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2035522
Reviewed-by: Owen Rodley <orodley@chromium.org>
Commit-Queue: Joey Scarr <jsca@chromium.org>
Cr-Commit-Position: refs/heads/master@{#28852}