[monorail] Add a parameterized selector for getting Issue by v1 API name.
The new v1 API returns only references to objects. E.g. GetHotlist()
only contains the API names of the owner and editors, and
ListHotlistItems() only contains the API names of the Issues.

So, we need to be able to fetch and access arbitrary Users and Issues from
the Redux store using the v1 API names.

I've chosen `issue.issue(name)` because we plan on making it the most
canonical selector with the v1 API (replacing selectors like
issueForRefString). The currently viewed issue is issue.viewedIssue().

Change-Id: I072101ed1e37b82c24a249f2ff1496438cab3b2b
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2056874
Commit-Queue: Dave Tu <dtu@chromium.org>
Reviewed-by: Tiffany Zhang <zhangtiff@chromium.org>
Cr-Commit-Position: refs/heads/master@{#29088}