[monorail] Add an action to fetch a list of issueRefs.
The v1 Hotlist API `ListHotlistItems` only contains references to the
hotlist issues. So we need a subsequent fetch to get all the Issue data.
`fetchIssues()` will be the call to do that.

There's significant overlap between this call and `fetchRelatedIssues()`
and `fetchIssueList()`, but if we're changing to a completely new API
anyway, it should be fine to just take on some tech debt here, rather
than trying to deduplicate them.

Change-Id: I948d8160358c1205fc8a3ac1b1cd864e7dfc2f02
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2056880
Commit-Queue: Dave Tu <dtu@chromium.org>
Reviewed-by: Tiffany Zhang <zhangtiff@chromium.org>
Reviewed-by: Kevin Weng <kweng@google.com>
Cr-Commit-Position: refs/heads/master@{#29120}