[🚝] Get Google Analytics for individual user pages.
Previously, we munged the location reported to GA to obscure user
emails. But that went a little too far -- we couldn't distinguish
between different user pages. So, only obscure the email from the
URL and nothing else.

We want this to get numbers on the popularity of Hotlist features.
For example: grid and chart view, grouping and sorting.

Change-Id: I9b0514c1e487550fbc2d0e7f16f587337de4bd37
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2081151
Reviewed-by: Tiffany Zhang <zhangtiff@chromium.org>
Reviewed-by: Kevin Weng <kweng@google.com>
Commit-Queue: Dave Tu <dtu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#29397}