cros_test_platform: Fail execute on errors
Before this CL, skylab-execute would fail with return code 2
on partial errors, to indicate that some response is available.
This led to errors being silently ignored in prod.

This CL changes the error code to 1, which will be interpreted
in prod as infrastructure failure.

Bug: 1028257
Change-Id: I8d3b07b864da218ab12eae71bc6f635bb5c4c538
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2083893
Reviewed-by: Xinan Lin <linxinan@chromium.org>
Commit-Queue: Prathmesh Prabhu <pprabhu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#29428}