cros lab inventory: a cron job to sync dev list to drone config
This cron job run daily to sync all device with the drone config entity
which provides the DUTs to be published to the drone queen.

This cron job is not suppose to do anything in normal if
add-dut/remove-dut works fine. But we still need it in case of data
inconsistency happens (e.g. manually remove some device entities, but
forget to clear them in the drone config).

BUG=None
TEST=`make test`
TEST=Ran on dev instance

Change-Id: I38c05d3ff0d337dfdac149667a222111907de783
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2087134
Commit-Queue: Congbin Guo <guocb@chromium.org>
Reviewed-by: Xixuan Wu <xixuan@chromium.org>
Auto-Submit: Congbin Guo <guocb@chromium.org>
Cr-Commit-Position: refs/heads/master@{#29455}