[🚝] Fix SPA hotlist rerank when items are filtered from the list.
We only pass in already filtered items to <mr-issue-list>, so when it
computes the target index of moved items, it doesn't include hidden
items. Instead of modifying <mr-issue-list> to be aware of them, we just
make <mr-hotlist-issues-page> sum up the hidden items to compute an
adjusted index.

Bug: monorail:6857
Change-Id: Ic3294b275b1f408e6354b1dedc75da2645cb8e4d
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2093372
Reviewed-by: Tiffany Zhang <zhangtiff@chromium.org>
Commit-Queue: Dave Tu <dtu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#29561}