cros_test_platform: Implement FetchResults
This CL is not reusing the bb client from the skylab tool because there
are major differences in business logic between the two (e.g. the fields
being fetched and proto format of the results) and the amount of logic
shared between the two clients is minimal.

Bug: 1033291
Change-Id: I9b828092ef4135c24c89d0fba714e1e5cf809a7b
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2091630
Reviewed-by: Prathmesh Prabhu <pprabhu@chromium.org>
Commit-Queue: Alex Zamorzaev <zamorzaev@chromium.org>
Cr-Commit-Position: refs/heads/master@{#29678}