[🚝] Add a shared button bar component.
The button bar holds list controls (Example: at the top of Gmail.)
We want a shared component because we will use it in the main issue
list as well as the hotlist issue list, with different buttons.
We might make it part of <mr-issue-list> in the future, but that's
already a very large file.

The Material design spec doesn't really talk about button bars, but
it has a similar looking concept called toggle buttons:
https://material.io/components/buttons/#toggle-button

The button bar contains actions can that apply to the entire list or
to just the selected items. My suggestion would be to only show the
"list-scoped" actions when nothing is selected, and the
"selection-scoped" actions when the list has stuff selected.

Change-Id: I09f6e9011fbf56b56dab96e64d3d2ae9b9daf52c
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2099622
Reviewed-by: Tiffany Zhang <zhangtiff@chromium.org>
Commit-Queue: Dave Tu <dtu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#29680}