Initial implementation of the Pinpoint Service
This is an incomplete implementation of the Pinpoint Service, which
we're developing as a gRPC service.

In this change, we are proxying requests to the ScheduleJob handler
towards the legacy Pinpoint API. We're testing the conversion from the
protobuf definition of the API towards the structure that the legacy API
form handler supports.

We're also baking in requirements for the Endpoint Service Proxy header
which will do the authorisation checks for the service. Testing this
will require a production deployment of the API, and a change in the
legacy API to support requests made by the service account used by the
service when deployed in production.

R=abennetts@google.com, fancl@chromium.org

Bug: chromium:1059654, chromium:1059667
Change-Id: I05aeebf96e37760851e0967d6a27a94deae8a5fe
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2094778
Commit-Queue: Dean Berris <dberris@chromium.org>
Reviewed-by: Nodir Turakulov <nodir@chromium.org>
Auto-Submit: Dean Berris <dberris@chromium.org>
Cr-Commit-Position: refs/heads/master@{#30101}