[recipes] Handle rollback in image builders correctly.
It's a regression added in
https://chromium.googlesource.com/infra/infra/+/53d7f6d

When a builder builds a "Revert ..." CL, it very often ends up
with a previously known image/tarball whose pin is already in
the infradata/{k8s|gae} repo.

In this case roll_{images|tarballs}.py returns empty "new pins added",
but still makes a change to `channels.json` to indicate that this older
existing image is a new "staging" now. This change should still be
committed to actually rollback the deployment.

R=iannucci@chromium.org, tandrii@chromium.org

Change-Id: I2f0bd29e62a6760dab8f5d1615623081b68e210c
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2136391
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>
Reviewed-by: Andrii Shyshkalov <tandrii@google.com>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Cr-Commit-Position: refs/heads/master@{#30338}